Written a introduction for our REST convention.
It has three parts, the first is an introduction which shows the history
of REST. The second part is the reason for having a REST convention. The
final part is a reading guide about what is in the chapters of the REST
convention.

Once the reference implementations are finished the introduction should
explain them a little bit as well.